pyproject.toml
[tool.poetry]
name = "test-task-studio-team"
version = "0.1.0"
description = ""
authors = ["Дима Бамбагаев <bambagaevdmitrij@gmail.com>"]
readme = "README.md"
packages = [{include = "test_task_studio_team"}]

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^2.25.1"
aioredis = "^2.0.1"
environs = "^9.5.0"
transliterate = "^1.10.2"
aiohttp = "^3.8.5"
pydantic = "^2.1.1"
pillow = "9.5.0"
django = "^4.2.3"
djangorestframework = "^3.14.0"
django-filter = "^23.2"
django-cors-headers = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
